Add WebMVC Controller test

+++ src/main/java/com/lucaskruk/todowebapp/services/TaskServiceImpl.java
@@ -1,5 +1,7 @@
 package com.lucaskruk.todowebapp.services;
 
+import com.lucaskruk.todowebapp.exceptions.InvalidRequestException;
+import com.lucaskruk.todowebapp.exceptions.NotFoundException;
 import com.lucaskruk.todowebapp.models.Task;
 import com.lucaskruk.todowebapp.repositories.TaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,26 +27,43 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public Task updateTask(int id, Task task) {
-        Task taskFromDB = taskRepository.findById(id).get();
-        taskFromDB.setTaskname(task.getTaskname());
-        taskFromDB.setDueDate(task.getDueDate());
-        taskFromDB.setStatus(task.getStatus());
-        return taskRepository.save(taskFromDB);
+
+        if (task == null) {
+            throw new InvalidRequestException("Task to update must not be null");
+        }
+
+        Optional<Task> taskFromDB = taskRepository.findById(id);
+
+        if (taskFromDB.isEmpty()) {
+            throw new NotFoundException("Task with id: " + id + "does not exists.");
+        }
+        Task foundTask = taskFromDB.get();
+        foundTask.setTaskname(task.getTaskname());
+        foundTask.setDueDate(task.getDueDate());
+        foundTask.setStatus(task.getStatus());
+        return taskRepository.save(foundTask);
     }
 
     @Override
     public Task getTask(int id) {
         Optional<Task> task = taskRepository.findById(id);
+        if (task.isEmpty()) {
+            throw new NotFoundException("Task with id: " + id + "does not exists.");
+        }
         return task.get();
     }
 
     @Override
     public List<Task> getAllTasks() {
-        return (List<Task>) taskRepository.findAll();
+        return taskRepository.findAll();
     }
 
     @Override
     public void deleteTask(int id) {
+        Optional<Task> task = taskRepository.findById(id);
+        if (task.isEmpty()) {
+            throw new NotFoundException("Task with id: " + id + "does not exists.");
+        }
         taskRepository.deleteById(id);
     }
 }
